feat(utilities): add function to create a cart item modifier summary
